Fix circular import: move types into `_utils` and rename `_utils` into `common`
